fs-1373: Added feature toggle for snapshot scheduling and timeline

+++ source/orch-mgr/configservice/configservice.cpp
@@ -309,8 +309,8 @@ class ConfigurationServiceHandler : virtual public ConfigurationServiceIf {
     }
 
     int64_t createSnapshotPolicy(boost::shared_ptr<fds::apis::SnapshotPolicy>& policy) {
-        if (configDB->createSnapshotPolicy(*policy)) {
-            om->snapshotMgr.addPolicy(*policy);
+        if (om->enableSnapshotSchedule && configDB->createSnapshotPolicy(*policy)) {
+            om->snapshotMgr->addPolicy(*policy);
             return policy->id;
         }
         return -1;
@@ -322,8 +322,10 @@ class ConfigurationServiceHandler : virtual public ConfigurationServiceIf {
     }
 
     void deleteSnapshotPolicy(boost::shared_ptr<int64_t>& id) {
-        configDB->deleteSnapshotPolicy(*id);
-        om->snapshotMgr.removePolicy(*id);
+        if (om->enableSnapshotSchedule) {
+            configDB->deleteSnapshotPolicy(*id);
+            om->snapshotMgr->removePolicy(*id);
+        }
     }
 
     void attachSnapshotPolicy(boost::shared_ptr<int64_t>& volumeId,
@@ -418,7 +420,7 @@ class ConfigurationServiceHandler : virtual public ConfigurationServiceIf {
             apiException("error creating volume");
         } else {
             // volume created successfully ,
-            // no create a base snapshot. [FS-471]
+            // now create a base snapshot. [FS-471]
             // we have to do this here because only OM can create a new
             // volume id
             boost::shared_ptr<int64_t> sp_volId(new int64_t(vol->rs_get_uuid().uuid_get_val()));
@@ -464,7 +466,9 @@ class ConfigurationServiceHandler : virtual public ConfigurationServiceIf {
             apiException(err.GetErrstr());
         }
         // add this snapshot to the retention manager ...
-        om->snapshotMgr.deleteScheduler->addSnapshot(snapshot);
+        if (om->enableSnapshotSchedule) {
+            om->snapshotMgr->deleteScheduler->addSnapshot(snapshot);
+        }
     }
 };
